remove encoding.ts, move serialization to provider.ts
