+++ src/providers/ark.ts
@@ -7,7 +7,9 @@ import {
     ArkInfo,
 } from "./base";
 import type { VirtualCoin } from "../types/wallet";
-import { VtxoTree } from "../core/vtxoTree";
+import { VtxoTree } from "../core/tree/vtxoTree";
+import { TreeNonces, TreePartialSigs } from "../core/signingSession";
+import { hex } from "@scure/base";
 
 // Define event types
 export interface ArkEvent {
@@ -314,7 +316,7 @@ export class ArkProvider extends BaseArkProvider {
     async submitTreeNonces(
         settlementID: string,
         pubkey: string,
-        nonces: string
+        nonces: TreeNonces
     ): Promise<void> {
         const url = `${this.serverUrl}/v1/round/tree/submitNonces`;
         const response = await fetch(url, {
@@ -325,7 +327,7 @@ export class ArkProvider extends BaseArkProvider {
             body: JSON.stringify({
                 roundId: settlementID,
                 pubkey,
-                treeNonces: nonces,
+                treeNonces: encodeNoncesMatrix(nonces),
             }),
         });
 
@@ -338,7 +340,7 @@ export class ArkProvider extends BaseArkProvider {
     async submitTreeSignatures(
         settlementID: string,
         pubkey: string,
-        signatures: string
+        signatures: TreePartialSigs
     ): Promise<void> {
         const url = `${this.serverUrl}/v1/round/tree/submitSignatures`;
         const response = await fetch(url, {
@@ -349,7 +351,7 @@ export class ArkProvider extends BaseArkProvider {
             body: JSON.stringify({
                 roundId: settlementID,
                 pubkey,
-                treeSignatures: signatures,
+                treeSignatures: encodeSignaturesMatrix(signatures),
             }),
         });
 
@@ -532,11 +534,108 @@ export class ArkProvider extends BaseArkProvider {
             return {
                 type: SettlementEventType.SigningNoncesGenerated,
                 id: data.roundSigningNoncesGenerated.id,
-                treeNonces: data.roundSigningNoncesGenerated.treeNonces,
+                treeNonces: decodeNoncesMatrix(
+                    hex.decode(data.roundSigningNoncesGenerated.treeNonces)
+                ),
             };
         }
 
         console.warn("Unknown event structure:", data);
         return null;
     }
+}
+
+function encodeMatrix(matrix: Uint8Array[][]): Uint8Array {
+    // Calculate total size needed:
+    // 4 bytes for number of rows
+    // For each row: 4 bytes for length + sum of encoded cell lengths
+    let totalSize = 4;
+    for (const row of matrix) {
+        totalSize += 4; // row length
+        for (const cell of row) {
+            totalSize += cell.length;
+        }
+    }
+
+    // Create buffer and DataView
+    const buffer = new ArrayBuffer(totalSize);
+    const view = new DataView(buffer);
+    let offset = 0;
+
+    // Write number of rows
+    view.setUint32(offset, matrix.length, true); // true for little-endian
+    offset += 4;
+
+    // Write each row
+    for (const row of matrix) {
+        // Write row length
+        view.setUint32(offset, row.length, true);
+        offset += 4;
+
+        // Write each cell
+        for (const cell of row) {
+            new Uint8Array(buffer).set(cell, offset);
+            offset += cell.length;
+        }
+    }
+
+    return new Uint8Array(buffer);
+}
+
+function decodeMatrix(matrix: Uint8Array, cellLength: number): Uint8Array[][] {
+    // Create DataView to read the buffer
+    const view = new DataView(
+        matrix.buffer,
+        matrix.byteOffset,
+        matrix.byteLength
+    );
+    let offset = 0;
+
+    // Read number of rows
+    const numRows = view.getUint32(offset, true); // true for little-endian
+    offset += 4;
+
+    // Initialize result matrix
+    const result: Uint8Array[][] = [];
+
+    // Read each row
+    for (let i = 0; i < numRows; i++) {
+        // Read row length
+        const rowLength = view.getUint32(offset, true);
+        offset += 4;
+
+        const row: Uint8Array[] = [];
+
+        // Read each cell in the row
+        for (let j = 0; j < rowLength; j++) {
+            const cell = new Uint8Array(
+                matrix.buffer,
+                matrix.byteOffset + offset,
+                cellLength
+            );
+            row.push(new Uint8Array(cell));
+            offset += cellLength;
+        }
+
+        result.push(row);
+    }
+
+    return result;
+}
+
+function decodeNoncesMatrix(matrix: Uint8Array): TreeNonces {
+    const decoded = decodeMatrix(matrix, 66);
+    return decoded.map((row) => row.map((nonce) => ({ pubNonce: nonce })));
+}
+
+function encodeNoncesMatrix(nonces: TreeNonces): string {
+    return hex.encode(
+        encodeMatrix(nonces.map((row) => row.map((nonce) => nonce.pubNonce)))
+    );
+}
+
+function encodeSignaturesMatrix(signatures: TreePartialSigs): string {
+    return hex.encode(
+        encodeMatrix(signatures.map((row) => row.map((s) => s.encode())))
+    );
 }
